+++ app/src/main/java/net/bunny/android/demo/livestream/LiveStreamEditorViewModel.kt
@@ -94,7 +94,7 @@ class LiveStreamEditorViewModel : ViewModel() {
     val uiState = mutableUiState.asStateFlow()
 
     private val libraryId: Long
-        get() = BunnyStreamApi.libraryId
+        get() = App.di.libraryId
 
     private val repository
         get() = App.di.streamSdk.liveStreamRepository